v1.04-20240715
Fixed the crash if default playlist is not active or not responding - playlists screen should now show instead.
Blue button functionality returns on the start menu - to delete the working json data files - Use when things are going weird.
Code tweak to how and when custom sids (satellite refs) for epgs are picked up.
Removed blank epg importer source file. Not needed. It gets built if required.
Translation updates for spain, china and germany.
New intro video selection option - for those users putting alternative videos in the video folder.
Slight tweak to the logic of determining if a playlist has been switched or not.